mcp_server/requirements.txt
azure-functions
pandas
requests
